move tests into new file
